requirements.txt
django
djangorestframework
psycopg2-binary
celery
redis
uwsgi
gunicorn
nox
